Fixing stylecop warnings Adding symbols to nuget
