+++ src/BuildIt.Synchronization/BuildIt.Synchronization.nuspec
@@ -10,17 +10,18 @@
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
     <projectUrl>https://github.com/builttoroam/BuildIt/</projectUrl>
     <description>This library contains a number of helper methods and functions that make building applications on a variety of platforms easier</description>
-    <copyright>Copyright © 2016</copyright>
-    <tags>BuildIt</tags>
+    <copyright>Copyright © 2017</copyright>
+    <tags>BuildIt, Synchronization</tags>
     <dependencies>      
-        <dependency id="BuildIt.General" version="1.0.0.41" />
-        <dependency id="NETStandard.Library" version="1.6.1" />
-        <dependency id="Newtonsoft.Json" version="8.0.3"/>
+        <dependency id="BuildIt.General" version="1.0.0.75" />
+        <dependency id="Newtonsoft.Json" version="10.0.3"/>
     </dependencies>
   </metadata>
   <files>
     <file src="BuildIt.Synchronization\bin\Release\BuildIt.Synchronization.dll" target="lib\netstandard1.0" />
     <file src="BuildIt.Synchronization\bin\Release\BuildIt.Synchronization.pdb" target="lib\netstandard1.0" />
     <file src="BuildIt.Synchronization\bin\Release\BuildIt.Synchronization.xml" target="lib\netstandard1.0" />
+
+    <file src="BuildIt.Synchronization\**\*.cs" target="src\BuildIt.Synchronization\BuildIt.Synchronization" exclude="BuildIt.Synchronization\obj\**\*.cs" />
   </files>
 </package>
