+++ src/BuildIt.Synchronization/BuildIt.Synchronization/SyncAction.cs
@@ -1,0 +1,12 @@
+namespace BuildIt.Synchronization
+{
+    public enum SyncAction
+    {
+        None,
+        Start,
+        Progress,
+        End,
+        Error,
+        Cancel
+    }
+}
