--- src/BuildIt.Synchronization/BuildIt.Synchronization/project.json
@@ -1,11 +1,0 @@
-﻿{
-  "supports": {},
-  "dependencies": {
-    "BuildIt.General": "1.0.0.41",
-    "NETStandard.Library": "1.6.1",
-    "Newtonsoft.Json": "9.0.1"
-  },
-  "frameworks": {
-    "netstandard1.0": {}
-  }
-}
